Move signing session id to sub package
