+++ relayer/session/signing/session_id.go
@@ -1,4 +1,4 @@
-package session
+package signing
 
 import (
 	"fmt"
